refactor(daemon): keep explain-request validation with the kernel authority; RPC layer stays shape-only
