+++ packages/kernel/test/contracts/entity-action-explanation.test.ts
@@ -5,6 +5,7 @@ import {
   ENTITY_ACTION_EXPLANATION_SCHEMA,
   getEntityKindContract,
   taskActionUsage,
+  validateEntityActionExplainRequest,
   validateEntityActionExplanationSet,
   type ActorIdentity,
   type EntityActionContract,
@@ -15,6 +16,34 @@ import {
 import { serializeEntityActionExplanationSet } from "../../src/domain/entity-action-explanation.ts";
 
 const actor: ActorIdentity = { principal: { personId: "person-explain" }, executor: null };
+
+test("Entity Action explain request is exact and caps object batches at 500 refs", () => {
+  assert.deepEqual(
+    validateEntityActionExplainRequest({
+      schema: "entity-action-explain-request/v1",
+      mode: "object",
+      refs: Array.from({ length: 500 }, () => "task/task-1"),
+    }),
+    [],
+  );
+  assert.match(
+    validateEntityActionExplainRequest({
+      schema: "entity-action-explain-request/v1",
+      mode: "object",
+      refs: Array.from({ length: 501 }, () => "task/task-1"),
+    }).join("\n"),
+    /1\.\.500/u,
+  );
+  assert.match(
+    validateEntityActionExplainRequest({
+      schema: "entity-action-explain-request/v1",
+      mode: "catalog",
+      refs: [],
+      actor,
+    }).join("\n"),
+    /unknown/u,
+  );
+});
 
 test("catalog and object modes share one strict schema with four honest criterion states", () => {
   const { ref, actions } = taskCatalog(),
